+++ test/xml/nodes/cdata_test.dart
@@ -18,6 +18,7 @@ void main() {
     expect(node.document, same(document));
     expect(node.depth, 2);
     expect(node.attributes, isEmpty);
+    expect(node.elementAttributes, isEmpty);
     expect(node.namespaces, isEmpty);
     expect(node.children, isEmpty);
     expect(node.nodeType, XmlNodeType.CDATA);
@@ -38,6 +39,7 @@ void main() {
     expect(node.document, same(document));
     expect(node.depth, 2);
     expect(node.attributes, isEmpty);
+    expect(node.elementAttributes, isEmpty);
     expect(node.namespaces, isEmpty);
     expect(node.children, isEmpty);
     expect(node.nodeType, XmlNodeType.CDATA);