+++ test/xml/nodes/comment_test.dart
@@ -16,6 +16,7 @@ void main() {
     expect(node.document, same(document));
     expect(node.depth, 2);
     expect(node.attributes, isEmpty);
+    expect(node.elementAttributes, isEmpty);
     expect(node.namespaces, isEmpty);
     expect(node.children, isEmpty);
     expect(node.nodeType, XmlNodeType.COMMENT);
@@ -33,6 +34,7 @@ void main() {
     expect(node.document, same(document));
     expect(node.depth, 2);
     expect(node.attributes, isEmpty);
+    expect(node.elementAttributes, isEmpty);
     expect(node.namespaces, isEmpty);
     expect(node.children, isEmpty);
     expect(node.nodeType, XmlNodeType.COMMENT);